requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
asyncpg==0.29.0
pydantic==2.7.1
pydantic[email]==2.7.1
python-dotenv==1.0.1
redis==5.0.7
aio-pika==9.4.1
